refactor: 상품 상세 페이지 구조 변경 및 스타일 파일 재구성
