+++ src/pages/Explore/ui/ProductDetail/ui/ProductItemDetail/ProductItemDetail.styled.ts
